3.0.9 webc plat 1762 float

+++ KalturaMediaServer/src/com/kaltura/media/server/managers/KalturaLiveManager.java
@@ -375,7 +375,8 @@ abstract public class KalturaLiveManager extends KalturaManager implements ILive
 				return;
 			}
 
-			updatedLiveEntry.currentBroadcastStartTime = (int) (new Date().getTime() / 1000);
+			// Set the time to be unix time stamp with milliseconds as the decimal fraction
+			updatedLiveEntry.currentBroadcastStartTime = new Date().getTime() / 1000.0;
 
 			KalturaClient impersonateClient = impersonate(liveEntry.partnerId);
 			try {
@@ -523,7 +524,7 @@ abstract public class KalturaLiveManager extends KalturaManager implements ILive
 		splitRecordingTimer.purge();
 	}
 	
-	public void appendRecording(String entryId, String assetId, KalturaMediaServerIndex index, String filePath, float duration) {
+	public void appendRecording(String entryId, String assetId, KalturaMediaServerIndex index, String filePath, double duration) {
 
 		logger.info("Entry [" + entryId + "] asset [" + assetId + "] index [" + index + "] filePath [" + filePath + "] duration [" + duration + "]");
 		
@@ -552,7 +553,7 @@ abstract public class KalturaLiveManager extends KalturaManager implements ILive
 		
 		try {
 			
-			Method method = liveServiceInstance.getClass().getMethod("appendRecording", String.class, String.class, KalturaMediaServerIndex.class, KalturaDataCenterContentResource.class, float.class);
+			Method method = liveServiceInstance.getClass().getMethod("appendRecording", String.class, String.class, KalturaMediaServerIndex.class, KalturaDataCenterContentResource.class, double.class);
 			KalturaLiveEntry updatedEntry = (KalturaLiveEntry)method.invoke(liveServiceInstance, entryId, assetId, index, resource, duration);
 			
 			if(updatedEntry != null){
@@ -677,7 +678,7 @@ abstract public class KalturaLiveManager extends KalturaManager implements ILive
 		return liveEntry;
 	}
 	
-	protected boolean saveUploadAsXml (String entryId, String assetId, KalturaMediaServerIndex index, String filePath, float duration, int partnerId)
+	protected boolean saveUploadAsXml (String entryId, String assetId, KalturaMediaServerIndex index, String filePath, double duration, int partnerId)
 	{
 		try {
 			DocumentBuilderFactory docFactory = DocumentBuilderFactory.newInstance();
@@ -708,7 +709,7 @@ abstract public class KalturaLiveManager extends KalturaManager implements ILive
 			
 			// duration element
 			Element durationElem = doc.createElement("duration");
-			durationElem.appendChild(doc.createTextNode(Float.toString(duration)));
+			durationElem.appendChild(doc.createTextNode(Double.toString(duration)));
 			rootElement.appendChild(durationElem);
 			
 			// filepath element
